backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
graphiti-core==0.28.2
neo4j==5.26.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.11.5
openai==1.91.0
python-multipart==0.0.20
trafilatura==2.0.0
